+++ dio/CHANGELOG.md
@@ -5,7 +5,9 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 ## Unreleased
 
-*None.*
+- Graceful handling of responses with nonzero `Content-Length`, `Content-Type` json, but empty body
+  - Empty responses are now transformed to `null`
+
 
 ## 5.6.0